+++ oshi-core/src/main/java/oshi/hardware/platform/windows/WindowsUsbDevice.java
@@ -21,26 +21,22 @@ package oshi.hardware.platform.windows;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.sun.jna.Native;
-import com.sun.jna.platform.win32.SetupApi;
-import com.sun.jna.platform.win32.SetupApi.SP_DEVINFO_DATA;
-import com.sun.jna.platform.win32.WinNT;
-import com.sun.jna.platform.win32.WinNT.HANDLE;
 import com.sun.jna.ptr.IntByReference;
-import com.sun.jna.ptr.NativeLongByReference;
 
 import oshi.hardware.UsbDevice;
 import oshi.hardware.common.AbstractUsbDevice;
-import oshi.hardware.platform.mac.MacUsbDevice;
 import oshi.jna.platform.windows.Cfgmgr32;
+import oshi.jna.platform.windows.Cfgmgr32Util;
 import oshi.util.MapUtil;
 import oshi.util.ParseUtil;
 import oshi.util.platform.windows.WmiUtil;
@@ -54,6 +50,32 @@ public class WindowsUsbDevice extends AbstractUsbDevice {
     private static final long serialVersionUID = 2L;
 
     private static final Logger LOG = LoggerFactory.getLogger(WindowsUsbDevice.class);
+
+    enum USBControllerProperty implements WmiProperty {
+        PNPDEVICEID(ValueType.STRING);
+
+        private ValueType type;
+
+        USBControllerProperty(ValueType type) {
+            this.type = type;
+        }
+
+        @Override
+        public ValueType getType() {
+            return this.type;
+        }
+    }
+
+    private static List<String> controllerDeviceIdList = new ArrayList<>();
+    static {
+        // One time lookup of USB Controller PnP Device IDs which don't change
+        WmiQuery<USBControllerProperty> usbControllerQuery = WmiUtil.createQuery("Win32_USBController",
+                USBControllerProperty.class);
+        WmiResult<USBControllerProperty> usbController = WmiUtil.queryWMI(usbControllerQuery);
+        for (int i = 0; i < usbController.getResultCount(); i++) {
+            controllerDeviceIdList.add((String) usbController.get(USBControllerProperty.PNPDEVICEID).get(i));
+        }
+    }
 
     enum PnPEntityProperty implements WmiProperty {
         NAME(ValueType.STRING), //
@@ -72,7 +94,8 @@ public class WindowsUsbDevice extends AbstractUsbDevice {
         }
     }
 
-    private static final WmiQuery<PnPEntityProperty> PNPENTITY_QUERY = WmiUtil.createQuery("Win32_PnPEntity",
+    private static final String PNPENTITY_BASE_CLASS = "Win32_PnPEntity";
+    private static final WmiQuery<PnPEntityProperty> PNPENTITY_QUERY = WmiUtil.createQuery(null,
             PnPEntityProperty.class);
 
     enum DiskDriveProperty implements WmiProperty {
@@ -91,39 +114,31 @@ public class WindowsUsbDevice extends AbstractUsbDevice {
         }
     }
 
-    private static final WmiQuery<DiskDriveProperty> DISKDRIVE_QUERY = WmiUtil.createQuery("Win32_DiskDrive",
+    private static final String DISKDRIVE_BASE_CLASS = "Win32_DiskDrive";
+    private static final WmiQuery<DiskDriveProperty> DISKDRIVE_QUERY = WmiUtil.createQuery(null,
             DiskDriveProperty.class);
-    private static final WmiQuery<DiskDriveProperty> PHYSICALMEDIA_QUERY = WmiUtil.createQuery("Win32_PhysicalMedia",
+    private static final String PHYSICALMEDIA_BASE_CLASS = "Win32_PhysicalMedia";
+    private static final WmiQuery<DiskDriveProperty> PHYSICALMEDIA_QUERY = WmiUtil.createQuery(null,
             DiskDriveProperty.class);
-
-    enum USBControllerProperty implements WmiProperty {
-        PNPDEVICEID(ValueType.STRING);
-
-        private ValueType type;
-
-        USBControllerProperty(ValueType type) {
-            this.type = type;
-        }
-
-        @Override
-        public ValueType getType() {
-            return this.type;
-        }
-    }
-
-    private static final WmiQuery<USBControllerProperty> USB_CONTROLLER_QUERY = WmiUtil
-            .createQuery("Win32_USBController", USBControllerProperty.class);
 
     private static final Pattern VENDOR_PRODUCT_ID = Pattern
             .compile(".*(?:VID|VEN)_(\\p{XDigit}{4})&(?:PID|DEV)_(\\p{XDigit}{4}).*");
 
     /*
-     * Maps to store information using PNPDeviceID as the key
+     * Map to store information using PNPDeviceID as the key
      */
-    private static Map<String, String> nameMap = new HashMap<>();
-    private static Map<String, String> vendorMap = new HashMap<>();
-    private static Map<String, String> serialMap = new HashMap<>();
-    private static Map<String, List<String>> hubMap = new HashMap<>();
+    private static Map<String, WindowsUsbDevice> usbDeviceCache = new HashMap<>();
+
+    /*
+     * Sets to track USB devices in cache
+     */
+    private static Set<String> devicesToAdd = new HashSet<>();
+    private static Set<String> devicesToRemove = null;
+
+    /*
+     * Map to build the recursive tree structure
+     */
+    private static Map<String, List<String>> deviceTreeMap = new HashMap<>();
 
     public WindowsUsbDevice(String name, String vendor, String vendorId, String productId, String serialNumber,
             UsbDevice[] connectedDevices) {
@@ -150,57 +165,91 @@ public class WindowsUsbDevice extends AbstractUsbDevice {
     private static void addDevicesToList(List<UsbDevice> deviceList, UsbDevice[] connectedDevices) {
         for (UsbDevice device : connectedDevices) {
             deviceList.add(new WindowsUsbDevice(device.getName(), device.getVendor(), device.getVendorId(),
-                    device.getProductId(), device.getSerialNumber(), new MacUsbDevice[0]));
+                    device.getProductId(), device.getSerialNumber(), new UsbDevice[0]));
             addDevicesToList(deviceList, device.getConnectedDevices());
         }
     }
 
     private static UsbDevice[] getUsbDevices() {
-        // Start by collecting information for all PNP devices. While in theory
-        // these could be individually queried with a WHERE clause, grabbing
-        // them all up front incurs minimal memory overhead in exchange for
-        // faster access later
+        // The controllerDeviceIDList contains the PnPDeviceIDs of the
+        // controllers.
 
-        // Clear maps
-        nameMap.clear();
-        vendorMap.clear();
-        serialMap.clear();
-        hubMap.clear();
+        // The deviceCache stores UsbDevices that we have already returned
+        // Navigate the device tree to see what's present: remove from cache if
+        // no longer present and add to cache if not present
 
-        // Query Win32_PnPEntity to populate the maps
-        WmiResult<PnPEntityProperty> pnpEntity = WmiUtil.queryWMI(PNPENTITY_QUERY);
-        for (int i = 0; i < pnpEntity.getResultCount(); i++) {
-            String pnpDeviceID = (String) pnpEntity.get(PnPEntityProperty.PNPDEVICEID).get(i);
-            nameMap.put(pnpDeviceID, (String) pnpEntity.get(PnPEntityProperty.NAME).get(i));
-            String vendor = (String) pnpEntity.get(PnPEntityProperty.MANUFACTURER).get(i);
-            if (!vendor.isEmpty()) {
-                vendorMap.put(pnpDeviceID, vendor);
-            }
-        }
+        // Clear recursive ID map and set up sets
+        deviceTreeMap.clear();
+        // Add any seen in tree that aren't in cache
+        devicesToAdd.clear();
+        // Also remove from cache if not seen in tree
+        devicesToRemove = new HashSet<String>(usbDeviceCache.keySet());
 
-        // Get serial # for disk drives or other physical media
-        WmiResult<DiskDriveProperty> serialNumber = WmiUtil.queryWMI(DISKDRIVE_QUERY);
-        for (int i = 0; i < serialNumber.getResultCount(); i++) {
-            serialMap.put((String) serialNumber.get(DiskDriveProperty.PNPDEVICEID).get(i),
-                    ParseUtil.hexStringToString((String) serialNumber.get(DiskDriveProperty.SERIALNUMBER).get(i)));
-        }
-        serialNumber = WmiUtil.queryWMI(PHYSICALMEDIA_QUERY);
-        for (int i = 0; i < serialNumber.getResultCount(); i++) {
-            serialMap.put((String) serialNumber.get(DiskDriveProperty.PNPDEVICEID).get(i),
-                    ParseUtil.hexStringToString((String) serialNumber.get(DiskDriveProperty.SERIALNUMBER).get(i)));
-        }
-
-        // Build the device tree. Start with the USB Controllers
-        // and recurse downward to devices as needed
-        WmiResult<USBControllerProperty> usbController = WmiUtil.queryWMI(USB_CONTROLLER_QUERY);
-
-        List<UsbDevice> controllerDevices = new ArrayList<>();
-        for (int i = 0; i < usbController.getResultCount(); i++) {
-            String controllerDeviceId = (String) usbController.get(USBControllerProperty.PNPDEVICEID).get(i);
+        List<WindowsUsbDevice> controllerDevices = new ArrayList<>();
+        for (String controllerDeviceId : controllerDeviceIdList) {
             putChildrenInDeviceTree(controllerDeviceId, 0);
+            updateDeviceCache();
             controllerDevices.add(getDeviceAndChildren(controllerDeviceId, "0000", "0000"));
         }
-        return controllerDevices.toArray(new UsbDevice[controllerDevices.size()]);
+        return controllerDevices.toArray(new WindowsUsbDevice[controllerDevices.size()]);
+
+    }
+
+    private static void updateDeviceCache() {
+
+        // Remove devices no longer in tree
+        for (String deviceID : devicesToRemove) {
+            usbDeviceCache.remove(deviceID);
+            LOG.debug("Removing {} from USB device cache.", deviceID);
+        }
+        devicesToRemove = null;
+        // Create list to add
+        if (!devicesToAdd.isEmpty()) {
+            StringBuilder sb = new StringBuilder();
+            boolean first = true;
+            for (String deviceID : devicesToAdd) {
+                if (first) {
+                    sb.append(" WHERE (PnPDeviceID=\"");
+                    first = false;
+                } else {
+                    sb.append(" OR (PnPDeviceID=\"");
+                }
+                sb.append(deviceID).append("\")");
+            }
+            String whereClause = sb.toString();
+            // Query Win32_PnPEntity to populate the maps
+            PNPENTITY_QUERY.setWmiClassName(PNPENTITY_BASE_CLASS + whereClause);
+            WmiResult<PnPEntityProperty> pnpEntity = WmiUtil.queryWMI(PNPENTITY_QUERY);
+            for (int i = 0; i < pnpEntity.getResultCount(); i++) {
+                String pnpDeviceID = (String) pnpEntity.get(PnPEntityProperty.PNPDEVICEID).get(i);
+                String name = (String) pnpEntity.get(PnPEntityProperty.NAME).get(i);
+                String vendor = (String) pnpEntity.get(PnPEntityProperty.MANUFACTURER).get(i);
+                WindowsUsbDevice device = new WindowsUsbDevice(name, vendor, null, null, "", new WindowsUsbDevice[0]);
+                usbDeviceCache.put(pnpDeviceID, device);
+                LOG.debug("Adding {} to USB device cache.", pnpDeviceID);
+            }
+            // Get serial # for disk drives or other physical media
+            DISKDRIVE_QUERY.setWmiClassName(DISKDRIVE_BASE_CLASS + whereClause);
+            WmiResult<DiskDriveProperty> serialNumber = WmiUtil.queryWMI(DISKDRIVE_QUERY);
+            for (int i = 0; i < serialNumber.getResultCount(); i++) {
+                String pnpDeviceID = (String) serialNumber.get(DiskDriveProperty.PNPDEVICEID).get(i);
+                if (usbDeviceCache.containsKey(pnpDeviceID)) {
+                    WindowsUsbDevice device = usbDeviceCache.get(pnpDeviceID);
+                    device.serialNumber = ParseUtil
+                            .hexStringToString((String) serialNumber.get(DiskDriveProperty.SERIALNUMBER).get(i));
+                }
+            }
+            PHYSICALMEDIA_QUERY.setWmiClassName(PHYSICALMEDIA_BASE_CLASS + whereClause);
+            serialNumber = WmiUtil.queryWMI(PHYSICALMEDIA_QUERY);
+            for (int i = 0; i < serialNumber.getResultCount(); i++) {
+                String pnpDeviceID = (String) serialNumber.get(DiskDriveProperty.PNPDEVICEID).get(i);
+                if (usbDeviceCache.containsKey(pnpDeviceID)) {
+                    WindowsUsbDevice device = usbDeviceCache.get(pnpDeviceID);
+                    device.serialNumber = ParseUtil
+                            .hexStringToString((String) serialNumber.get(DiskDriveProperty.SERIALNUMBER).get(i));
+                }
+            }
+        }
     }
 
     /**
@@ -214,74 +263,38 @@ public class WindowsUsbDevice extends AbstractUsbDevice {
      *            the code will search for a match.
      */
     private static void putChildrenInDeviceTree(String deviceId, int deviceInstance) {
-        int devInst = deviceInstance;
-        // If no devInst provided, find it by matching deviceId
-        if (devInst == 0) {
-            // Get a handle to the device with this deviceId
-            // Start with all classes
-            HANDLE hinfoSet = SetupApi.INSTANCE.SetupDiGetClassDevs(null, null, null, SetupApi.DIGCF_ALLCLASSES);
-            if (hinfoSet == WinNT.INVALID_HANDLE_VALUE) {
-                LOG.error("Invalid handle value for {}. Error code: {}", deviceId, Native.getLastError());
-                return;
-            }
-            // Iterate to find matching parent
-            SP_DEVINFO_DATA dinfo = new SP_DEVINFO_DATA();
-            dinfo.cbSize = dinfo.size();
-            int i = 0;
-            while (SetupApi.INSTANCE.SetupDiEnumDeviceInfo(hinfoSet, i++, dinfo)) {
-                if (deviceId.equals(getDeviceId(dinfo.DevInst))) {
-                    devInst = dinfo.DevInst;
-                    break;
-                }
-            }
+        // Track deviceIDs to add/remove from cache
+        devicesToRemove.remove(deviceId);
+        if (!usbDeviceCache.containsKey(deviceId)) {
+            devicesToAdd.add(deviceId);
         }
+        // If no devInst provided, find it
+        int devInst = deviceInstance;
         if (devInst == 0) {
-            LOG.error("Unable to find a devnode handle for {}.", deviceId);
-            return;
+            IntByReference pdnDevInst = new IntByReference();
+            Cfgmgr32.INSTANCE.CM_Locate_DevNode(pdnDevInst, deviceId, 0);
+            devInst = pdnDevInst.getValue();
         }
         // Now iterate the children. Call CM_Get_Child to get first child
         IntByReference child = new IntByReference();
         if (0 == Cfgmgr32.INSTANCE.CM_Get_Child(child, devInst, 0)) {
             // Add first child to a list
             List<String> childList = new ArrayList<>();
-            String childId = getDeviceId(child.getValue());
+            String childId = Cfgmgr32Util.CM_Get_Device_ID(child.getValue());
             childList.add(childId);
-            hubMap.put(deviceId, childList);
+            deviceTreeMap.put(deviceId, childList);
             putChildrenInDeviceTree(childId, child.getValue());
             // Find any other children
             IntByReference sibling = new IntByReference();
             while (0 == Cfgmgr32.INSTANCE.CM_Get_Sibling(sibling, child.getValue(), 0)) {
                 // Add to the list
-                String siblingId = getDeviceId(sibling.getValue());
-                hubMap.get(deviceId).add(siblingId);
+                String siblingId = Cfgmgr32Util.CM_Get_Device_ID(sibling.getValue());
+                deviceTreeMap.get(deviceId).add(siblingId);
                 putChildrenInDeviceTree(siblingId, sibling.getValue());
                 // Make this sibling the new child to find other siblings
                 child = sibling;
             }
         }
-    }
-
-    /**
-     * Gets the device id for a devnode
-     *
-     * @param devInst
-     *            the handle to the devnode
-     * @return The PNPDeviceID
-     */
-    private static String getDeviceId(int devInst) {
-        NativeLongByReference ulLen = new NativeLongByReference();
-        if (0 != Cfgmgr32.INSTANCE.CM_Get_Device_ID_Size(ulLen, devInst, 0)) {
-            LOG.error("Couldn't get device string for device instance {}", devInst);
-            return "";
-        }
-        // Add 1 for null terminator
-        int size = ulLen.getValue().intValue() + 1;
-        char[] buffer = new char[size];
-        if (0 != Cfgmgr32.INSTANCE.CM_Get_Device_ID(devInst, buffer, size, 0)) {
-            LOG.error("Couldn't get device string for device instance {} with size {}", devInst, size);
-            return "";
-        }
-        return new String(buffer).trim();
     }
 
     /**
@@ -304,14 +317,19 @@ public class WindowsUsbDevice extends AbstractUsbDevice {
             vendorId = m.group(1).toLowerCase();
             productId = m.group(2).toLowerCase();
         }
-        List<String> pnpDeviceIds = MapUtil.getOrDefault(hubMap, hubDeviceId, new ArrayList<String>());
+        List<String> pnpDeviceIds = MapUtil.getOrDefault(deviceTreeMap, hubDeviceId, new ArrayList<String>());
         List<WindowsUsbDevice> usbDevices = new ArrayList<>();
         for (String pnpDeviceId : pnpDeviceIds) {
             usbDevices.add(getDeviceAndChildren(pnpDeviceId, vendorId, productId));
         }
         Collections.sort(usbDevices);
-        return new WindowsUsbDevice(MapUtil.getOrDefault(nameMap, hubDeviceId, vendorId + ":" + productId),
-                MapUtil.getOrDefault(vendorMap, hubDeviceId, ""), vendorId, productId,
-                MapUtil.getOrDefault(serialMap, hubDeviceId, ""), usbDevices.toArray(new UsbDevice[usbDevices.size()]));
+        WindowsUsbDevice device = usbDeviceCache.get(hubDeviceId);
+        if (device.name.isEmpty()) {
+            device.name = vendorId + ":" + productId;
+        }
+        device.vendorId = vendorId;
+        device.productId = productId;
+        device.connectedDevices = usbDevices.toArray(new WindowsUsbDevice[usbDevices.size()]);
+        return device;
     }
 }